// src/modules/invoice/usecase/find-invoice/find-invoice.dto.ts
export interface FindInvoiceUsecaseInputDto {
    id: string;
}

export interface FindInvoiceUsecaseOutputDto {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: string;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
    createdAt: Date;
}